+++ src/org/deltik/mc/signedit/subcommands/SetSignSubcommand.java
@@ -20,33 +20,29 @@
 package org.deltik.mc.signedit.subcommands;
 
 import org.deltik.mc.signedit.ArgParser;
-import org.deltik.mc.signedit.ChatComms;
 import org.deltik.mc.signedit.SignText;
-import org.deltik.mc.signedit.SignTextHistoryManager;
 import org.deltik.mc.signedit.exceptions.LineSelectionException;
 import org.deltik.mc.signedit.exceptions.MissingLineSelectionException;
-import org.deltik.mc.signedit.interactions.SetSignEditInteraction;
 import org.deltik.mc.signedit.interactions.SignEditInteraction;
 
 import javax.inject.Inject;
+import javax.inject.Provider;
+import java.util.Map;
 
 public class SetSignSubcommand implements SignSubcommand {
+    private final Map<String, Provider<SignEditInteraction>> interactions;
     private final ArgParser argParser;
     private final SignText signText;
-    private final ChatComms comms;
-    private final SignTextHistoryManager historyManager;
 
     @Inject
     public SetSignSubcommand(
+            Map<String, Provider<SignEditInteraction>> interactions,
             ArgParser argParser,
-            SignText signText,
-            ChatComms comms,
-            SignTextHistoryManager historyManager
+            SignText signText
     ) {
+        this.interactions = interactions;
         this.argParser = argParser;
         this.signText = signText;
-        this.comms = comms;
-        this.historyManager = historyManager;
     }
 
     @Override
@@ -71,6 +67,6 @@ public class SetSignSubcommand implements SignSubcommand {
             signText.setLine(selectedLine, text);
         }
 
-        return new SetSignEditInteraction(signText, comms, historyManager);
+        return interactions.get("Set").get();
     }
 }